export functions, detect repeats
