+++ package.json
@@ -25,8 +25,10 @@
     "@popperjs/core": "^2.5.4",
     "react": "^17.0.1",
     "react-beautiful-dnd": "^13.0.0",
+    "react-bootstrap": "^1.4.0",
     "react-dom": "^17.0.1",
-    "react-modal": "^3.11.2"
+    "react-modal": "^3.11.2",
+    "react-popper": "^2.2.4"
   },
   "devDependencies": {
     "@babel/core": "^7.12.3",
