+++ src/components/app/Library/Library.jsx
@@ -5,53 +5,7 @@ import { DragDropContext, Droppable, Draggable } from 'react-beautiful-dnd';
 import { getBooks } from '../../../services/books-api';
 import ReactModal from 'react-modal';
 import Book from './Book';
-
-
-// a little function to help us with reordering the result
-const reorder = (list, startIndex, endIndex) => {
-  const result = Array.from(list);
-  const [removed] = result.splice(startIndex, 1);
-  result.splice(endIndex, 0, removed);
-
-  return result;
-};
-
-/*
- * Moves an item from one list to another list.
- */
-const move = (source, destination, droppableSource, droppableDestination) => {
-  const sourceClone = Array.from(source);
-  const destClone = Array.from(destination);
-  const [removed] = sourceClone.splice(droppableSource.index, 1);
-
-  destClone.splice(droppableDestination.index, 0, removed);
-  const result = {};
-  result[droppableSource.droppableId] = sourceClone;
-  result[droppableDestination.droppableId] = destClone;
-
-  return result;
-};
-
-const grid = 8;
-
-const getItemStyle = (isDragging, draggableStyle) => ({
-  // some basic styles to make the items look a bit nicer
-  userSelect: 'none',
-  padding: grid * 2,
-  margin: `0 0 ${grid}px 0`,
-
-  // change background colour if dragging
-  background: isDragging ? 'lightgreen' : 'grey',
-
-  // styles we need to apply on draggables
-  ...draggableStyle
-});
-
-const getListStyle = isDraggingOver => ({
-  background: isDraggingOver ? 'lightblue' : 'lightgrey',
-  padding: grid,
-  width: 250
-});
+import { move, reorder, getItemStyle, getListStyle } from '../../../utils/drag-functions';
 
 export default class Library extends Component {
   constructor() {
@@ -62,10 +16,11 @@ export default class Library extends Component {
       selectedItem: '',
       showModal: false,
       search: '',
-      searchResults: []
+      searchResults: [],
+      show: false
     };
     this.handleOpenModal = this.handleOpenModal.bind(this);
-    this.handleCloseModal = this.handleCloseModal.bind(this);
+    this.handleCloseModal = this.handleCloseModal.bind(this); 
   }
 
   handleOpenModal() {
@@ -138,10 +93,15 @@ export default class Library extends Component {
 
     addToList = (book) => {
       const items = this.state.items;
-      items.push(book);
-      this.setState({ items });
-      this.handleCloseModal();
-    }
+      if(book === items.find(x => x === book) || this.state.selected.find(x => x === book) === book) {
+        alert('REPEAT BOOK');
+      }
+      else {      
+        items.push(book);
+        this.setState({ items });
+        this.handleCloseModal();
+      }
+    };
 
     alertItem = (draggableId) => {
       const selectedItem = this.state.selected.find(x => x.id === draggableId);
